fix(proxy): forwarder child-count race + split pure core into its own TU
Second roundtable pass — two blocking items:
- aimee-forwarder: block SIGCHLD across fork()+g_children++ so the reaper cannot
  decrement for a child that exits before it is counted (which leaked the counter to
  MAX_CHILDREN and then permanently refused new tunnels — the very exhaustion the cap
  guards). The child restores the mask immediately after fork.
- Split the pure, side-effect-free decision core (classify/ip_is_blocked/port_allowed/
  host_allowed/default_allowlist) into sandbox_pkg_proxy_core.c so the unit test links
  it WITHOUT the logging/socket object (previously it only linked by luck of
  --gc-sections dropping the unused I/O). unit-test-sandbox-pkg-proxy now links the core.
Also (non-blocking): simplified proxy_forward_head's request-line rewrite to take the
validated 8-char HTTP version and always append CRLF (dropped the fragile CR trim).
